Refactor frontend to be more object-oriented     -fix method access in templates
