+++ website/addons/s3/templates/s3_credentials_modal.mako
@@ -33,10 +33,10 @@
 
                 <div class="modal-footer">
 
-                    <a href="#" class="btn btn-default" data-bind="click: options.clearModal" data-dismiss="modal">Cancel</a>
+                    <a href="#" class="btn btn-default" data-bind="click: clearModal" data-dismiss="modal">Cancel</a>
 
                     <!-- Save Button -->
-                    <button data-bind="click: options.connectAccount" class="btn btn-success">Save</button>
+                    <button data-bind="click: connectAccount" class="btn btn-success">Save</button>
 
                 </div><!-- end modal-footer -->
